% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clause.R
\name{add_clause}
\alias{add_clause}
\title{Add a new clause to the current contract}
\usage{
add_clause(contract, object)
}
\arguments{
\item{contract}{A [contract()].}

\item{object}{A clause object.}
}
\value{
A updated [contract()] with the new term in the last slot.
}
\description{
`add_clause` adds a step to the last location in the recipe.
}
