% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterparty.R
\name{add_counterparty}
\alias{add_counterparty}
\title{Add a counterparty to a contract}
\usage{
add_counterparty(contract, id)
}
\arguments{
\item{contract}{A contract to which the counterparty needs to be specified.}

\item{id}{The name or id of the counterparty.}
}
\description{
This function adds a counterparty to a tidy contract.
}
\examples{
library(magrittr)

make_contract() \%>\%
  add_counterparty("JP Morgan")

}
\keyword{contract,}
\keyword{counterparty}
