% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_terms.R
\name{add_date_term}
\alias{add_date_term}
\title{Add a new date term to the current contract}
\usage{
add_date_term(contract, date_type, date_value)
}
\arguments{
\item{contract}{A [contract()].}

\item{date_type}{The type of date to be added to the contract.}

\item{date_value}{The date of the event.}
}
\value{
A updated [contract()] with the new term in the last slot.
}
\description{
`add_date_term` specifies a date term to be added to a contract.
This is most useful when made part of a custom clause.
}
