% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_events.R
\name{add_events}
\alias{add_events}
\title{Add a generic financial event or set of events to a dataframe of financial events}
\usage{
add_events(df, evt_date, evt_type, evt_value, evt_currency)
}
\arguments{
\item{df}{a dataframe to which the event is to be added.}

\item{evt_currency}{The currency of the financial events.}

\item{evt_dates}{A date or vector of dates corresponding to the days on which the financial event or events will take place.}

\item{evt_types}{A string or vector of strings corresponding to the types of financial events taking place.}

\item{evt_values}{A numeric value or vector corresponding to the values of the financial events. Typically a currency value.}
}
\description{
This function adds a financial event or a set of financial events to a dataframe of financial events.
}
\examples{
make_date_seq(from=1,to=5, by=1)
}
\keyword{event}
