% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_maturity.R
\name{add_maturity}
\alias{add_maturity}
\title{Add maturity date to a contract}
\usage{
add_maturity(contract, date)
}
\arguments{
\item{contract}{A contract object}

\item{date}{The maturity date of the contract.}
}
\description{
Add maturity date to a contract
}
