% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stipulation.R
\name{add_stipulation}
\alias{add_stipulation}
\title{Overall wrapper for adding a new term to the current contract}
\usage{
add_stipulation(contract, stipulation, stipulation_name)
}
\arguments{
\item{contract}{A [contract()].}

\item{object}{A stipulation to add to the contract.}
}
\value{
A updated [contract()] with the new term in the last slot.
}
\description{
`add_stipulation` adds a term to the last location in the contract.
}
