% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{add_term}
\alias{add_term}
\title{Overall wrapper for adding a new term to the current contract}
\usage{
add_term(contract, term, term_name)
}
\arguments{
\item{contract}{A [contract()].}

\item{object}{A term to add to the contract.}
}
\value{
A updated [contract()] with the new term in the last slot.
}
\description{
`add_term` adds a term to the last location in the contract.
}
