% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contractr-package.R
\docType{package}
\name{contractr-package}
\alias{contractr}
\alias{contractr-package}
\title{contractr: contractr is a grammar of financial contracts.}
\description{
contractr is a grammar of financial contracts implemented in R. It can be used to compose financial contracts of different types.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jameslairdsmith/contractr}
  \item Report bugs at \url{https://github.com/jameslairdsmith/contractr/issues}
}

}
\author{
\strong{Maintainer}: James Laird-Smith \email{jameslairdsmith@gmail.com}

}
\keyword{internal}
