% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft.R
\name{draft}
\alias{draft}
\title{Fills in the missing parts of a contract}
\usage{
draft(contract)
}
\arguments{
\item{contract}{A contract object}
}
\description{
This function takes a contract object and replaces all the unknown elements that it can.
}
