% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_up.R
\name{draw_up}
\alias{draw_up}
\title{Fills in the missing parts of a contract}
\usage{
draw_up(contract)
}
\arguments{
\item{contract}{A contract object}
}
\description{
This function takes a contract object and replaces all the unknown elements that it can.
}
