% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter_contract.R
\name{enter_contract}
\alias{enter_contract}
\title{Create a contract type}
\usage{
enter_contract(starting)
}
\arguments{
\item{starting}{The date on which the contract begins.}

\item{exchanging}{Optional. The initial amount exchanged at the start of the contract. If omitted, defaults to 0.}
}
\description{
This function begins the process of creating a new contract type by adding an initial exchange date.
}
\examples{
library(lubridate)
enter_contract(on = dmy(01/01/2018))
}
\keyword{contract,}
\keyword{event}
