% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contract_start.R
\name{get_start_date}
\alias{get_start_date}
\title{Get the start date of a contract}
\usage{
get_start_date(df, alt_name = "Contract Deal Date")
}
\arguments{
\item{df}{A date to start from.}
}
\description{
This function gets the start date of a contract.
}
\keyword{date}
