% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initially_exchanging.R
\name{initially_exchanging}
\alias{initially_exchanging}
\title{Add initial exchange event to a contract}
\usage{
initially_exchanging(df, evt_value, evt_currency = "USD",
  on = "initial")
}
\arguments{
\item{df}{A contract to which the initial exchange is to be added.}

\item{evt_value}{The amount exchanged at the begining of the contract.}

\item{evt_currency}{The currency of the exchange amount. Defaults to "USD".}

\item{on}{The date on which this payment is made. Defaults to "initial", which gets the start of the contract already specified.}
}
\description{
This function adds an initial exchange event between the parties to a contact.
}
\examples{
library(lubridate)
enter_contract(on = dmy(01/01/2018))
}
\keyword{contract,}
\keyword{event}
