% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clause.R
\name{make_clause}
\alias{make_clause}
\title{Overall wrapper to make new clause objects}
\usage{
make_clause(subclass, ..., .suffix = "_clause")
}
\arguments{
\item{subclass}{A character string for the resulting class. For example,
if `subclass = "blah"` the term object that is returned has class
`clause_blah`.}

\item{...}{All arguments to the operator that should be returned.}

\item{.prefix}{Prefix to the subclass created.}
}
\value{
An updated term with the new class.
}
\description{
The function creates objects with class of `clause`
}
\keyword{contract}
\keyword{term}
