% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_date_seq.R
\name{make_date_seq}
\alias{make_date_seq}
\title{Create a sequence of dates}
\usage{
make_date_seq(start_date, end_date, period_gap, tolerance = 0.2,
  include_first = F, include_last = T, replace_last = T)
}
\arguments{
\item{start_date}{The start date}

\item{end_date}{The end date}

\item{period_gap}{The period gap}

\item{tolerance}{the tolerance}

\item{include_first}{whether to include the first date}

\item{include_last}{whether to include the last date}

\item{replace_last}{whether to replace the last date with \code{end_date}}
}
\description{
It is often the case that one needs to create
a series of dates using only the start, end and
freqency.
}
