% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{make_term}
\alias{make_term}
\title{Overall wrapper to make new term objects}
\usage{
make_term(subclass, ..., .prefix = "term_")
}
\arguments{
\item{subclass}{A character string for the resulting class. For example,
if \code{subclass = "blah"} the term object that is returned has class
\code{clause_blah}.}

\item{...}{All arguments to the operator that should be returned.}

\item{.prefix}{Prefix to the subclass created.}
}
\value{
An updated term with the new class.
}
\description{
The function creates objects with class of \code{term}
}
\keyword{contract}
\keyword{term}
