% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectors.R
\name{maturity}
\alias{maturity}
\alias{inception}
\title{Select properties of a contract}
\usage{
maturity(x)

inception(x)
}
\description{
This function allows users a quick shorthand for specifying the
  properties of a contract. These do not need to be specified in prior steps.
  Whether they are specified in future or prior steps, the values will only be replaced
  once the contract is sheduled.
}
\keyword{date}
