% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contract.R
\name{print.contract}
\alias{print.contract}
\title{Print a Contract}
\usage{
\method{print}{contract}(x, form_width = 30, ...)
}
\arguments{
\item{x}{A \code{contract} object}

\item{form_width}{The number of characters used to print}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The original object (invisibly)
}
\description{
Print a Contract
}
\author{
James Laird-Smith
}
