% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_check.R
\name{ellipse_check}
\alias{ellipse_check}
\title{Check for Empty Ellipses}
\usage{
ellipse_check(...)
}
\arguments{
\item{...}{Arguments pass in from a call to `step`}
}
\value{
If not empty, a list of quosures. If empty, an error is thrown.
}
\description{
Check for Empty Ellipses
}
\keyword{internal}
