% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule}
\alias{schedule}
\alias{schedule.contract}
\title{Create the events of a contract}
\usage{
schedule(object, ...)

\method{schedule}{contract}(contract)
}
\arguments{
\item{contract}{A contract with at least one term.}
}
\value{
A tibble housing the contract's events.
}
\description{
For a contract (with at least one term and no missing stipulations)
create the financial events that ought result from it.
}
