% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/someday.R
\name{someday_between}
\alias{someday_between}
\alias{someday}
\title{Specify unknown date values in a contract}
\usage{
someday_between(start_date, end_date)

someday()
}
\arguments{
\item{start_date}{the earliest possible date of the event.}

\item{end_date}{the latest possible date of the event.}
}
\description{
This function acts as a stand-in for unknown date values in a contract.
}
