% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starting_from.R
\name{starting_from}
\alias{starting_from}
\title{Add a start date to a contract}
\usage{
starting_from(contract, start_date, id = "inception")
}
\arguments{
\item{contract}{A contract for which the start date is to be specified.}

\item{date}{The start date}
}
\description{
This function specifies the start date to a tidy contract.
}
\examples{
library(magrittr)
library(lubridate)

make_contract() \%>\%
  starting_from(dmy("01/01/2000"))

}
\keyword{contract,}
\keyword{inception}
