% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stipulation.R
\name{stipulate}
\alias{stipulate}
\title{Add a stipulation to a contract}
\usage{
stipulate(contract, stipulation_value, stipulation_name)
}
\arguments{
\item{contract}{The \code{contract} to which the stipulation is to be added.}

\item{stipulation_value}{The value of the stipulation.}

\item{stipulation_name}{The name of the stipulation.}
}
\value{
A updated \code{contract} with the new stipulation in the last slot.
}
\description{
The function allows for the adding of a stipulation
to the list of stipulations in a contract. This is most
useful when used as a wrapper together with \code{add_term()}
and \code{make_term()} for creating a custom term type.
}
