% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{term}
\alias{term}
\title{Overall wrapper to make new term objects}
\usage{
term(subclass, ..., .prefix = "term_")
}
\arguments{
\item{subclass}{A character string for the resulting class. For example,
if `subclass = "blah"` the term object that is returned has class
`term_blah`.}

\item{...}{All arguments to the operator that should be returned.}

\item{.prefix}{Prefix to the subclass created.}
}
\value{
An updated term with the new class.
}
\description{
The function creates objects with class of `term`
}
\keyword{contract}
\keyword{term}
