% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end.R
\name{term_end}
\alias{term_end}
\title{Set the end details of a contract}
\usage{
term_end(contract, end_date, event_name = "end_date", payment = 0,
  receipt = 0)
}
\arguments{
\item{contract}{The contract for which the \code{end date} is to be specified.}

\item{end_date}{The end date to be specified.}

\item{event_name}{An optional name for the event. Defaults to "end_date"
(which is reccomended).}

\item{payment}{The amount paid at the end of the contract.
Defaults to 0.}

\item{receipt}{The amount recieved at the end of the contract.
Defaults to 0.}
}
\value{
A contract object with the term added.
}
\description{
This function can be used to specify the end date of a contract.
}
\details{
Note that \code{receipt} and \code{payment} cannot both be specified. If they are
an error will be thrown.
}
\examples{
library(magrittr)

contract_a <-
  contract() \%>\%
  term_end(as.Date("2000/01/01"))

contract_a

contract_a \%>\% schedule()

contract() \%>\%
 term_end(as.Date("2000/01/01"), event_name = "sale date") \%>\%
 schedule()

contract() \%>\%
 term_end(as.Date("2000/02/01"), receipt = 1000) \%>\%
 schedule()

contract() \%>\%
 term_end(as.Date("2000/03/01"), payment = 50) \%>\%
 schedule()
}
\keyword{contract,}
\keyword{date}
\keyword{end,}
