% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_series.R
\name{term_series}
\alias{term_series}
\title{Create a series of payments or reciepts}
\usage{
term_series(contract, freq, from, to, amount, event_name = "series", ...)
}
\arguments{
\item{contract}{The contract for which the \code{start date} is to be specified.}

\item{event_name}{An optional name for the event. Defaults to "start_date"
(which is reccomended).}

\item{start_date}{The start date to be specified.}

\item{payment}{The amount paid at the start of the contract.
Defaults to 0.}

\item{receipt}{The amount recieved at the start of the contract.
Defaults to 0.}
}
\value{
A contract object with the term added.
}
\description{
Create a series of payments or reciepts
}
\examples{
library(magrittr)
}
\keyword{contract,}
\keyword{date}
\keyword{start,}
