% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{term_start}
\alias{term_start}
\title{Set the start date of a contract}
\usage{
term_start(contract, start_date, id = "start_date", payment = 0,
  receipt = 0)
}
\arguments{
\item{contract}{The contract for which the start date is to be specified.}

\item{start_date}{The start date to be specified.}

\item{id}{An optional name for the event. Defaults to "start_date"
(which is reccomended).}

\item{payment}{The amount paid at the start of the contract.
Defaults to zero.}

\item{receipt}{The amount recieved at the start of the contract.
Defaults to zero.}
}
\value{
A contract object with the term added.
}
\description{
This function can be used to specify the start date of a contract.
}
\details{
If both `payment` and `receipt` are specified,
only the netted amount will be captured in the event.
}
\examples{
library(magrittr)

contract_a <-
  contract() \%>\%
  term_start(as.Date("2000/01/01"))

contract_a

contract_a \%>\% schedule()

contract() \%>\%
 term_start(as.Date("2000/02/01"), receipt = 100) \%>\%
 schedule()

contract() \%>\%
 term_start(as.Date("2000/03/01"), payment = 50) \%>\%
 schedule()

contract() \%>\%
 term_start(as.Date("2000/04/01"), payment = 50, receipt = 100) \%>\%
 schedule()
}
\keyword{contract,}
\keyword{date}
\keyword{start,}
