% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_date_seq.R
\name{translate_period_freq}
\alias{translate_period_freq}
\title{Modify periods into dates}
\usage{
translate_period_freq(period_string)
}
\arguments{
\item{period_string}{A string indicating the period.}
}
\description{
This function changes period strings into date strings readable by `seq()``
}
\examples{

translate_period_freq("quarterly")


}
\keyword{date}
\keyword{period,}
\keyword{string,}
