% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterparty.R
\name{with_counterparty}
\alias{with_counterparty}
\alias{check_counterparty}
\title{Add a counterparty to a contract.}
\usage{
with_counterparty(contract, counterparty_id)

check_counterparty(counterparty)
}
\arguments{
\item{contract}{A contract to which the counterparty needs to be specified.}

\item{counterparty_id}{The name or id of the counterparty.}
}
\description{
This function adds a counterparty to a tidy contract.
}
\keyword{contract,}
\keyword{counterparty}
