% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_maturity.R
\name{add_maturity}
\alias{add_maturity}
\title{Add a maturity date to a contract}
\usage{
add_maturity(contract, date)
}
\arguments{
\item{contract}{A contract to which the maturity date needs to be specified.}

\item{id}{The maturity date}
}
\description{
This function specifies the maturity date to a tidy contract.
}
\examples{
library(magrittr)

make_contract() \%>\%
  add_counterparty("JP Morgan")

}
\keyword{contract,}
\keyword{maturity}
