% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{add_term}
\alias{add_term}
\title{Add a new term contract}
\usage{
add_term(contract, term)
}
\arguments{
\item{contract}{A \code{\link[=contract]{contract()}}.}

\item{object}{A term object.}
}
\value{
A updated \code{\link[=contract]{contract()}} with the new term in the last slot.
}
\description{
\code{add_term} adds a term to the last location in the contract.
The function is most useful as a wrapper for the creation of
custom terms (in conjuction with \code{make_term()}.
}
