% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{add_term}
\alias{add_term}
\title{Add a new term to the current contract}
\usage{
add_term(contract, term)
}
\arguments{
\item{contract}{A \code{\link[=contract]{contract()}}.}

\item{object}{A term object.}
}
\value{
A updated \code{\link[=contract]{contract()}} with the new term in the last slot.
}
\description{
\code{add_term} adds a step to the last location in the recipe.
}
