% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{add_term}
\alias{add_term}
\title{Add a new term to the current contract}
\usage{
add_term(contract, object)
}
\arguments{
\item{contract}{A [contract()].}

\item{object}{A term object.}
}
\value{
A updated [contract()] with the new term in the last slot.
}
\description{
`add_term` adds a step to the last location in the recipe.
}
