% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contract.R
\name{contract}
\alias{contract}
\title{Create a tidy financial contract object}
\usage{
contract()
}
\description{
This function creates a financial contract object to which contractual terms can be added.
}
\keyword{contract}
