% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contractr-package.R
\docType{package}
\name{contractr-package}
\alias{contractr}
\alias{contractr-package}
\title{contractr: a grammar of financial contracts}
\description{
The `contractr` package can be used to compose financial contracts
  in R. Superficially, it may appear that financial contracts exist
  in huge variety, but in reality almost all are combinations of a
  few event patterns. `contractr` has built-in functions for each of
  these patterns, any of which can be added to contract as a "term"
  of that contract. Thus:
}
\details{
\enumerate{
\item A contract consists of one or more terms that determine the
 events that result from the contract.
\item Term can be added  to a contact sequentially using pipes.
}
}
\section{Basic Functions}{

The three main functions in contractr are [contract()], [draft()],
  and [schedule()].

[contract()] creates an empty contract object. to which terms
  can be added. [draft()] fills in the missing elements of a
  contract given the information already added to the contract
  or provided separately. Lastly the [schedule()] function returns
  the events that result from the contract.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jameslairdsmith/contractr}
  \item Report bugs at \url{https://github.com/jameslairdsmith/contractr/issues}
}

}
\author{
\strong{Maintainer}: James Laird-Smith \email{jameslairdsmith@gmail.com}

}
\keyword{internal}
