% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_date_seq.R
\name{make_date_seq}
\alias{make_date_seq}
\title{Create a series of dates}
\usage{
make_date_seq(...)
}
\arguments{
\item{from}{A date to start from}

\item{to}{A date to go to.}

\item{by}{spacing between dates.}
}
\description{
This function creates a series of dates evenly spread apart. Currently a thin wrapper around `by()`
}
\examples{
make_date_seq(from=1,to=5, by=1)
}
\keyword{date}
