% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule}
\alias{schedule}
\alias{schedule.contract}
\title{Create the events from a contract}
\usage{
schedule(object, ...)

\method{schedule}{contract}(object)
}
\arguments{
\item{object}{A contract with at least one term.}
}
\value{
A tibble
}
\description{
For a contract with at least one term and no missing stipulations
create the financial events that ought result from it.
}
\author{
James Laird-Smith
}
\keyword{contract}
\keyword{event}
