% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{term_start}
\alias{term_start}
\title{Add a start date to a contract}
\usage{
term_start(contract, start_date, id = "start_date")
}
\arguments{
\item{contract}{A contract for which the start date is to be specified.}

\item{date}{The start date}
}
\description{
This function specifies the start date to a tidy contract.
}
\examples{
library(magrittr)
library(lubridate)

make_contract() \%>\%
  term_start(dmy("01/01/2000"))

}
\keyword{contract,}
\keyword{inception}
