% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{term_start}
\alias{term_start}
\title{Set the start details of a contract}
\usage{
term_start(contract, start_date, id = "start_date", payment = 0,
  receipt = 0, consideration)
}
\arguments{
\item{contract}{The contract for which the \code{start date} is to be specified.}

\item{start_date}{The start date to be specified.}

\item{id}{An optional name for the event. Defaults to "start_date"
(which is reccomended).}

\item{payment}{The amount paid at the start of the contract.
Defaults to 0.}

\item{receipt}{The amount recieved at the start of the contract.
Defaults to 0.}
}
\value{
A contract object with the term added.
}
\description{
This function can be used to specify the start date of a contract.
}
\details{
Note that if both \code{receipt} and \code{payment} are specified, the net amount
(\code{receipt} - \code{payment}) will be the amount shown in the event.
}
\examples{
library(magrittr)

contract_a <-
  contract() \%>\%
  term_start(as.Date("2000/01/01"))

contract_a

contract_a \%>\% schedule()

contract() \%>\%
 term_start(as.Date("2000/01/01"), id = "purchase date") \%>\%
 schedule()

contract() \%>\%
 term_start(as.Date("2000/02/01"), receipt = 1000) \%>\%
 schedule()

contract() \%>\%
 term_start(as.Date("2000/03/01"), payment = 50) \%>\%
 schedule()

contract() \%>\%
 term_start(as.Date("2000/04/01"), payment = 50, receipt = 1000) \%>\%
 schedule()
}
\keyword{contract,}
\keyword{date}
\keyword{start,}
