% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{term_start}
\alias{term_start}
\title{Set the start date of a contract}
\usage{
term_start(contract, start_date, id = "start_date")
}
\arguments{
\item{contract}{The contract for which the start date is to be specified.}

\item{start_date}{The start date to be specified.}

\item{id}{An optional name for the event. Defaults to "start_date"
(which is reccomended).}

\item{payment}{The amount paid at the start of the contract.
Defaults to zero.}

\item{receipt}{The amount recieved at the start of the contract.
Defaults to zero.}
}
\description{
This function can be used to specify the start date of a contract.
}
\examples{
library(magrittr)

a_contract <-
  contract() \%>\%
  term_start(as.Date("2000/01/01"))

a_contract

a_contract \%>\% schedule()
}
\keyword{contract,}
\keyword{date}
\keyword{start,}
